requirements.txt
streamlit==1.31.1
langchain==0.1.9
qdrant-client==1.9.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==0.21.0
streamlit==1.31.1
langchain-google-vertexai==0.1.0